% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxtrackerR.R
\name{fx_conversion}
\alias{fx_conversion}
\title{Converts a specific amount of money from current currency (curr1) to desired currency (curr2)}
\usage{
fx_conversion(curr1, curr2, amt)
}
\arguments{
\item{curr1}{A character vector which is the ticker of the current currency such as 'EUR'}

\item{curr2}{A character vector which is the ticker of the desired currency to convert to such as 'USD'}

\item{amt}{A double vector that is the amount of money to be converted}
}
\value{
A double vector. Amount of value converted from currency 1 to currency 2
}
\description{
Converts a specific amount of money from current currency (curr1) to desired currency (curr2)
}
\examples{
fx_conversion('EUR', 'USD', 150.75)
}
