% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_rate_lookup.R
\name{fx_rate_lookup}
\alias{fx_rate_lookup}
\title{Return the first date (reverse chronological order) on which the target price falling between day high and day low based on the availability of data.}
\usage{
fx_rate_lookup(curr, target_px)
}
\arguments{
\item{curr}{A character vector which is the ticker of the currency pair such as 'EURUSD'}

\item{target_px}{A double vector that is the target price for the lookup}
}
\value{
A datetime. The closest date on which the target price falling between day high and day low.
}
\description{
Return the first date (reverse chronological order) on which the target price falling between day high and day low based on the availability of data.
}
\examples{
fx_rate_lookup('EURUSD', 1.072)
}
