% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_trend_viz.R
\name{price_trend_viz}
\alias{price_trend_viz}
\title{Visualizes trend of the exchange rate of a currency pair between the selected start date and end date.}
\usage{
price_trend_viz(curr, start_date, end_date, option)
}
\arguments{
\item{curr}{A character vector which is the ticker of the currency pair such as 'EURUSD'}

\item{start_date}{A character vector which is the start date of the selected period of time}

\item{end_date}{A character vector which is the end date of the selected period of time}

\item{option}{A character vector which must be from 'Open', High', "Low', 'Close'}
}
\value{
A line plot that shows the trend of the exchange rate of a currency pair over the selected period of time.
}
\description{
Visualizes trend of the exchange rate of a currency pair between the selected start date and end date.
}
\examples{
price_trend_viz('EURUSD', '2018-12-31', '2022-12-31')
}
